test css commands and modify home link into button
This commit has a lot off add ins which I am aware goes against the
idea of atomic commits but has no particular rhyme or reason besides
just testing out the uses of css and how certain aspects interact
resulting in most of the changes simply being spontaneous decisions.

Added in more divs to make it easier to work with css, and made the
return to home link into a button.